requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
redis>=4.6.0
firebase-admin==6.5.0
google-cloud-firestore==2.21.0
google-cloud-secret-manager==2.24.0
google-api-python-client==2.149.0
google-api-core==2.21.0
google-auth==2.40.3
grpcio==1.74.0
openai==1.97.1
anthropic==0.69.0
tiktoken>=0.12.0
tenacity==9.1.2
telethon==1.42.0
pydantic>=2.11.0
python-dotenv==1.1.1
# Data processing
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
# Celery pour les tâches parallèles
celery[redis]>=5.3.0
# Stripe pour les paiements
stripe==11.6.0
# Dépendances pour Google Drive et Sheets
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
# Dépendances pour l'interface utilisateur
colorama==0.4.6
# Multipart form data pour FastAPI UploadFile
python-multipart>=0.0.9
Pillow>=10.0.0
# Dépendances pour la conversion PDF en images
pdf2image>=1.16.0
# Dépendances pour le traitement des dates
pytz==2024.2
croniter==6.0.0
# Dépendances pour Google GenAI
google-generativeai>=0.3.0
google-genai==0.3.0
aiohttp==3.13.1
# Google Cloud Pub/Sub pour transport multi-canal (Telegram, Google Chat)
google-cloud-pubsub>=2.0.0
# AWS SDK pour CloudWatch Logs
boto3>=1.35.0
# PostgreSQL async driver pour Neon (module HR)
asyncpg>=0.29.0
# HTTP async client pour Jobber HR
httpx==0.27.0
# pytest-asyncio pour tests asynchrones
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-anyio>=0.0.0